+++ python/lsst/ts/wep/ButlerWrapper.py
@@ -75,12 +75,7 @@ class ButlerWrapper(object):
             Raw exposure object.
         """
 
-        # Data Id keys in w_2019_24:
-        # {'run': str, 'raftName': str, 'visit': int, 'detectorName': str,
-        #  'detector': int, 'snap': int}
-
         dataId = self._getDefaultDataId(visit, raft, sensor)
-        self._extendDataId(dataId, snap=snap)
 
         return self._butler.get("raw", dataId=dataId)
 
@@ -102,29 +97,9 @@ class ButlerWrapper(object):
             Default data Id.
         """
 
-        dataId = dict(visit=int(visit), raftName=raft, detectorName=sensor)
+        dataId = dict(expId=int(visit), raftName=raft, detectorName=sensor)
 
         return dataId
-
-    def _extendDataId(self, dataId, snap=None, aFilter=None):
-        """Extend the data Id.
-
-        Parameters
-        ----------
-        dataId : dict
-            Data Id.
-        snap : int, optional
-            Snap time (0 or 1) means first/ second exposure. (the default is
-            None.)
-        aFilter : str, optional
-            Active filter ("u", "g", "r", "i", "z", "y") (the default is None.)
-        """
-
-        if (snap is not None) and isinstance(snap, (int, float)):
-            dataId["snap"] = int(snap)
-
-        if (aFilter is not None) and isinstance(aFilter, str):
-            dataId["filter"] = aFilter
 
     def getPostIsrCcd(self, visit, raft, sensor, aFilter=None):
         """Get the post-ISR CCD exposure.
@@ -149,14 +124,30 @@ class ButlerWrapper(object):
             Post-ISR CCD object.
         """
 
-        # Data Id keys in w_2019_24:
-        # {'visit': int, 'filter': str, 'raftName': str, 'detectorName': str,
-        #  'detector': int}
-
         dataId = self._getDefaultDataId(visit, raft, sensor)
         self._extendDataId(dataId, aFilter=aFilter)
 
         return self._butler.get("postISRCCD", dataId=dataId)
+
+    def _extendDataId(self, dataId, snap=None, aFilter=None):
+        """Extend the data Id.
+
+        Parameters
+        ----------
+        dataId : dict
+            Data Id.
+        snap : int, optional
+            Snap time (0 or 1) means first/ second exposure. (the default is
+            None.)
+        aFilter : str, optional
+            Active filter ("u", "g", "r", "i", "z", "y") (the default is None.)
+        """
+
+        if (snap is not None) and isinstance(snap, (int, float)):
+            dataId["snap"] = int(snap)
+
+        if (aFilter is not None) and isinstance(aFilter, str):
+            dataId["filter"] = aFilter
 
     def getEimage(self, visit, raft, sensor, snap=None):
         """Get the PhoSim eimage exposure.
@@ -179,8 +170,6 @@ class ButlerWrapper(object):
             Eimage exposure object.
         """
 
-        # Data Id keys in w_2019_24:
-        # {'visit': int, 'snap': int, 'raftName': str, 'detectorName': str}
         dataId = self._getDefaultDataId(visit, raft, sensor)
         self._extendDataId(dataId, snap=snap)